+++ log-elasticsearch/CHANGELOG.md
@@ -1,3 +1,7 @@
+# log-elasticsearch-0.8.1 (2016-03-27)
+* Log.Backend.ElasticSearch.Internal now exports 'EsUsername' and
+  'EsPassword'.
+
 # log-elasticsearch-0.8 (2016-03-16)
 * Made ElasticSearchConfig an abstract type (#27).
 * Added support for HTTPS and basic auth (#26).